+++ pymince/patterns.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+class Singleton(type):
+    """Metaclass that makes a class a Singleton."""
+
+    _instances = dict()
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super().__call__(*args, **kwargs)
+        return cls._instances[cls]