+++ tests/test_patterns/test_singleton.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+import pymince.patterns
+
+
+class MyClass(metaclass=pymince.patterns.Singleton):
+    pass
+
+
+def test_singleton():
+    assert MyClass() is MyClass()